🚧 VMPWJP acr: split generation module
